Add `fill` TessellateIPU primitive mapping popops vertex.
Compared to the `full` LAX operation, the `fill_p` TessellateIPU primitive
ensures that the full array is allocated and filled with a given value (and not broadcasted).

As an useful example of using `fill`, in combination with `scatter`, we implement
`tile_sharded_identity` which creates an identity matrix, with proper tile mapping, without
introducing additional communications.