IPU Jacobi eigh `fori_loop` transition
This PR is switching the IPU Jacobi `eigh` implementation from a
Python loop to a JAX `fori_loop`, reducing massively code size and
allowing Eigen decomposition on larger matrices.

Note: in terms of performance, some issues still remain as Poplar
compiler is adding some extra on-tile copies which in theory could
be eluded.